refactor: move the lint result cluster to homeboy-extension-contract
The finale of the phase-subsystem breakdown - lint is the fourth and last
(after bench, trace, test). Moves the lint result types + the top-level
LintCommandOutput aggregate wholesale:
- StreamCaptureMetadata (core stream_capture.rs), SelfCheckCaptureMetadata,
  LintSummaryOutput, FormattingFindings - all pure serde
- LintCommandOutput aggregate - its field deps all resolve to contract/leaf
  crates: CiContext/PhaseReport/PhaseFailure (contract), FindingProducerSummary/
  HomeboyFinding (finding), AppliedRefactor (refactor-contract),
  BaselineComparison (= engine-primitives baseline::Comparison alias)

The lint run/report builder functions stay in core. Core re-exports everything.
Lint reverse-edge surface collapses to 1 core file - matching bench (58->4),
trace (45->3), test (->1). All four extension phase subsystems are now fully
backed by contract types.
